backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pydantic==2.10.3
sqlalchemy==2.0.23
alembic==1.13.0
gunicorn==21.2.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
black==23.12.1
flake8==6.1.0
mypy==1.8.0
